+++ src/main/java/com/example/bcsd/repository/ArticleRepository.java
@@ -1,48 +1,13 @@
 package com.example.bcsd.repository;
 
 import com.example.bcsd.model.Article;
-import org.springframework.stereotype.Repository;
 
-import java.util.*;
+import java.util.List;
 
-@Repository
-public class ArticleRepository {
-    private Long key = 1L;
-    private final Map<Long, Article> articles = new HashMap<>();
-
-    public List<Article> findAll() {
-        List<Article> result = new ArrayList<>();
-
-        articles.forEach((id, article) ->
-                result.add(article));
-
-        return result;
-    }
-
-    public Article findById(Long id) {
-        Article article = articles.get(id);
-
-        if (article == null) {
-            throw new NoSuchElementException("해당 게시글을 찾을 수 없습니다.");
-        }
-
-        return article;
-    }
-
-    public Article save(Article article) {
-        article.setId(key);
-        articles.put(key++, article);
-
-        return article;
-    }
-
-    public Article save(Long id, Article article) {
-        articles.put(id, article);
-
-        return article;
-    }
-
-    public void delete(Long id) {
-        articles.remove(id);
-    }
+public interface ArticleRepository {
+    List<Article> findAll();
+    Article findById(Long id);
+    Article save(Article article);
+    Article save(Long id, Article article);
+    void delete(Long id);
 }
