+++ src/main/java/com/example/bcsd/service/ArticleService.java
@@ -4,7 +4,7 @@ import com.example.bcsd.dto.request.ArticleCreateRequest;
 import com.example.bcsd.dto.resopnse.ArticleResponse;
 import com.example.bcsd.dto.request.ArticleUpdateRequest;
 import com.example.bcsd.model.Article;
-import com.example.bcsd.repository.ArticleRepository;
+import com.example.bcsd.repository.ArticleRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ArticleService {
-    private final ArticleRepository articleRepository;
+    private final ArticleRepositoryImpl articleRepository;
 
     public List<ArticleResponse> getAllArticles() {
         return articleRepository.findAll()
