+++ src/main/java/com/example/bcsd/service/ArticleViewService.java
@@ -2,8 +2,8 @@ package com.example.bcsd.service;
 
 import com.example.bcsd.dto.resopnse.ArticleViewResponse;
 import com.example.bcsd.model.Article;
-import com.example.bcsd.repository.ArticleRepository;
-import com.example.bcsd.repository.MemberRepository;
+import com.example.bcsd.repository.ArticleRepositoryImpl;
+import com.example.bcsd.repository.MemberRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,8 +12,8 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ArticleViewService {
-    private final ArticleRepository articleRepository;
-    private final MemberRepository memberRepository;
+    private final ArticleRepositoryImpl articleRepository;
+    private final MemberRepositoryImpl memberRepository;
 
     public List<ArticleViewResponse> getAllPostViews() {
         List<Article> articles = articleRepository.findAll();
