+++ src/main/java/com/example/bcsd/service/MemberService.java
@@ -3,7 +3,7 @@ package com.example.bcsd.service;
 import com.example.bcsd.dto.request.MemberRequest;
 import com.example.bcsd.dto.resopnse.MemberResponse;
 import com.example.bcsd.model.Member;
-import com.example.bcsd.repository.MemberRepository;
+import com.example.bcsd.repository.MemberRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +12,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class MemberService {
-    private final MemberRepository memberRepository;
+    private final MemberRepositoryImpl memberRepository;
 
     public List<MemberResponse> getAllMembers() {
         return memberRepository.findAll()
